% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC.R
\name{consensus.cluster}
\alias{consensus.cluster}
\title{Title}
\usage{
consensus.cluster(
  x,
  k.min = 3,
  k.max = 100,
  n.reps = 100,
  p.item = 0.8,
  p.feature = 1,
  dist.method = "euclidean",
  linkage = "complete",
  upper.lim = 0.9,
  lower.lim = 0.1,
  calculate.cdf = FALSE,
  cdf.length = 1000,
  calculate.dist.upfront = FALSE
)
}
\arguments{
\item{lower.lim}{}
}
\value{

}
\description{
Title
}
