% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{create_clustering,numeric-method}
\alias{create_clustering,numeric-method}
\title{Title}
\usage{
\S4method{create_clustering}{numeric}(clustering_result, alpha = 0.9)
}
\arguments{
\item{clustering_result}{A numeric vector of cluster labels for each element.}
}
\value{
A Clustering object.
}
\description{
Title
}
