% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{create_monocle_object}
\alias{create_monocle_object}
\title{Create monocle object based on expression matrix}
\usage{
create_monocle_object(
  normalized_expression_matrix,
  count_matrix = NULL,
  clustassess_object,
  metadata,
  stable_feature_type,
  stable_feature_set_size,
  stable_clustering_method,
  stable_n_clusters = NULL,
  gene_names = NULL,
  cell_names = NULL,
  is_object_from_shinyapp = FALSE
)
}
\description{
to be completed
}
