% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{element.sim.elscore}
\alias{element.sim.elscore}
\title{Title}
\usage{
element.sim.elscore(
  clustering1,
  clustering2,
  alpha = 0.9,
  r = 1,
  r2 = NULL,
  rescale_path_type = "max",
  relabeled_elements = NULL,
  ppr_implementation = "prpack"
)
}
\arguments{
\item{ppr_implementation}{}
}
\value{

}
\description{
Title
}
