% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{element_frustration}
\alias{element_frustration}
\title{Title}
\usage{
element_frustration(clustering_list)
}
\arguments{
\item{clustering_list}{A list of Clustering objects used to calculate
the element-wise frustration.}
}
\value{
a vector containing the element-wise frustration.
}
\description{
Title
}
