% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{element_sim}
\alias{element_sim}
\title{The element-centric clustering similarity.}
\usage{
element_sim(clustering1, clustering2)
}
\arguments{
\item{clustering1}{The first Clustering.}

\item{clustering2}{The second Clustering.}
}
\value{
The element-wise similarity between the two Clusterings.
}
\description{
The element-centric clustering similarity.
}
