% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{length,Clustering-method}
\alias{length,Clustering-method}
\title{Length of an Object}
\usage{
\S4method{length}{Clustering}(x)
}
\arguments{
\item{x}{The Clustering object.}
}
\value{
The number of elements.
}
\description{
Get the number of elements in the Clustering.
}
\examples{
km.res = kmeans(mtcars, 3)$cluster
km.clustering = create_clustering(km.res)
length(km.clustering)
}
