% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{merge_partitions}
\alias{merge_partitions}
\title{Merge Partitions}
\usage{
merge_partitions(
  partition_list,
  ecs_thresh = 1,
  order = TRUE,
  check_ties = FALSE
)
}
\arguments{
\item{partition_list}{A list of flat disjoint membership vectors.}

\item{ecs_thresh}{A numeric: the ecs threshold.}

\item{order}{A logical: if TRUE, order the partitions based on their frequencies.}

\item{check_ties}{A logical value that indicates whether to check for ties
in the highest frequency partitions or not. If TRUE, the function will put
at the first position the partition that has the highest similarity
with the other partitions. Defaults to \code{FALSE}.}
}
\value{
a list of the merged partitions
}
\description{
Merge flat disjoint clusterings whose pairwise ECS score is
above a given threshold. The merging is done using a complete linkage approach.
}
\examples{
initial_list <- list(c(1, 1, 2), c(2, 2, 2), c("B", "B", "A"))
merge_partitions(initial_list, 1)
}
