% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC.R
\name{pac_convergence}
\alias{pac_convergence}
\title{PAC Convergence Plot}
\usage{
pac_convergence(pac_res, k_plot)
}
\arguments{
\item{pac_res}{The data.frame output by consensus_cluster.}

\item{k_plot}{A vector with values of k to plot.}
}
\value{
A ggplot2 object with the convergence plot. Convergence has been
reached when the lines flatten out across k_plot values.
out across
}
\description{
Plot PAC across iterations for a set of k to assess convergence.
}
\examples{
pac.res = consensus_cluster(iris[,1:4], k_max=20)
pac_convergence(pac.res, k_plot=c(3,5,7,9))
}
