% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC.R
\name{pac_landscape}
\alias{pac_landscape}
\title{PAC Landscape Plot}
\usage{
pac_landscape(pac_res, n_shade = max(pac_res$iteration)/5)
}
\arguments{
\item{n_shade}{}
}
\value{

}
\description{
Plot final PAC values across range of k to find optimal number
of clusters.
}
