% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{plot_clustering_difference_facet}
\alias{plot_clustering_difference_facet}
\title{Graphical representation of the clustering method importance}
\usage{
plot_clustering_difference_facet(
  clustering_difference_object,
  embedding,
  low_limit = 0,
  high_limit = 1,
  grid = TRUE
)
}
\arguments{
\item{clustering_difference_object}{An object returned by the
`get_clustering_difference_object` method.}

\item{embedding}{An embedding (only the first two dimensions will be used for
visualisation).}

\item{low_limit}{The lowest value of ECC that will be displayed on the embedding.}

\item{high_limit}{The highest value of ECC that will be displayed on the embedding.}

\item{grid}{Boolean value indicating whether the facet should be a grid (where each
row is associated with a resolution value and each column with a clustering method) or
a wrap.}
}
\value{
A ggplot facet object
}
\description{
Display, for each clustering method and each resolution value,
the distribution of the EC consistency on a given embedding. We will use the `all`
field of the object returned by the `get_clustering_difference_object` method.
}
\examples{
adj_matrix = Seurat::FindNeighbors(as.matrix(mtcars),
    k.param = 5,
    nn.method = "rann",
    verbose = FALSE,
    compute.SNN = FALSE)$nn
clust_diff_obj = get_clustering_difference(graph_adjacency_matrix = adj_matrix,
    resolution = c(0.5, 1),
    n_repetitions = 1,
    algorithm = 1,
    verbose = FALSE)
plot_clustering_difference_facet(clust_diff_obj, as.matrix(mtcars[,1:2]))
}
