% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{plot_feature_stability_ecs_incremental}
\alias{plot_feature_stability_ecs_incremental}
\title{Feature Stability Incremental Boxplot}
\usage{
plot_feature_stability_ecs_incremental(
  feature_object_list,
  dodge_width = 0.7,
  text_size = 4
)
}
\arguments{
\item{feature_object_list}{An object or a concatenation of objects returned by the
`get_feature_stability_object` method.}

\item{dodge_width}{Used for adjusting the horizontal position of the boxplot; the value
will be passed in the `width` argument of the `ggplot2::position_dodge` method.}

\item{text_size}{The size of the labels above boxplots.}
}
\value{
A ggplot object
}
\description{
Perform an incremental ECS between two consecutive steps. The
ECS distribution will be displayed as a boxplot. Above each boxplot there will be
a pair of numbers representing the two steps that are compared.
}
\examples{
feature_stability_result = get_feature_stability(data_matrix = t(as.matrix(mtcars)),
   feature_set = colnames(mtcars),
   feature_type = "feature_name",
   steps = c(6,7),
   npcs = 2,
   n_repetitions = 1,
   algorithm = 1,
   # the following parameters are used by the umap function and are not mandatory
   n_neighbors = 2,
   approx_pow = TRUE,
   n_epochs = 0,
   init = "random",
   min_dist = 0.3)
plot_feature_stability_ecs_incremental(feature_stability_result, 0.7)
}
