% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{plot_feature_stability_mb_facet}
\alias{plot_feature_stability_mb_facet}
\title{Feature Stability - Cluster Membership facet}
\usage{
plot_feature_stability_mb_facet(
  feature_object_list,
  text_size = 5,
  n_facet_cols = 3,
  point_size = 0.3
)
}
\arguments{
\item{feature_object_list}{An object or a concatenation of objects returned by the
`get_feature_stability_object` method}

\item{text_size}{The size of the cluster label}

\item{n_facet_cols}{The number of facet's columns.}

\item{point_size}{The size of the points displayed on the plot.}
}
\value{
A ggplot facet object
}
\description{
Display a facet of plots where each graph is associated with a step of
a feature set and illustrates the distribution of the most frequent partition over the UMAP
embedding.
}
\examples{
feature_stability_result = get_feature_stability(data_matrix = t(as.matrix(mtcars)),
   feature_set = colnames(mtcars),
   feature_type = "feature_name",
   steps = -1,
   npcs = 2,
   n_repetitions = 1,
   algorithm = 1,
   # the following parameters are used by the umap function and are not mandatory
   n_neighbors = 3,
   approx_pow = TRUE,
   n_epochs = 0,
   init = "random",
   min_dist = 0.3)
plot_feature_stability_mb_facet(feature_stability_result)
}
