% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{print,Clustering-method}
\alias{print,Clustering-method}
\title{Print an Object}
\usage{
\S4method{print}{Clustering}(x)
}
\arguments{
\item{x}{The Clustering object.}
}
\description{
Prints out information about the Clustering, including
number of elements.
}
\examples{
km.res = kmeans(mtcars, 3)$cluster
km.clustering = create_clustering(km.res)
print(km.clustering)
}
