% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-feature-stability.R
\name{server_dimensionality_reduction}
\alias{server_dimensionality_reduction}
\title{Dimensionality reduction - server side}
\usage{
server_dimensionality_reduction(id, parent_session)
}
\description{
to be completed
}
