% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-graph-clustering.R
\name{server_graph_clustering}
\alias{server_graph_clustering}
\title{Server - Graph clustering module}
\usage{
server_graph_clustering(id, feature_choice, parent_session)
}
\arguments{
\item{id}{The id of the module, used to acess the UI elements.}

\item{feature_choice}{A reactive object that contains the chosen configuration
from the Dimensionality Reduction tab.}

\item{parent_session}{The session of the parent module, used to control the
tabs of the application.}
}
\description{
Creates the backend interface for the graph clustering module
inside the ClustAssess Shiny application.
}
\note{
This function should not be called directly, but in the context of the
app that is created using the \code{write_shiny_app} function.
}
