% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-feature-stability.R
\name{ui_dimensionality_reduction}
\alias{ui_dimensionality_reduction}
\title{UI - Dimensionality reduction module}
\usage{
ui_dimensionality_reduction(id)
}
\arguments{
\item{id}{The id of the module, used to identify the UI elements.}
}
\description{
Creates the UI interface for the dimensionality reduction
module inside the ClustAssess Shiny application.
}
\note{
This function should not be called directly, but in the context of the
app that is created using the \code{write_shiny_app} function.
}
