% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-landing-page.R
\name{ui_landing_page}
\alias{ui_landing_page}
\title{UI - Landing page module}
\usage{
ui_landing_page(id)
}
\arguments{
\item{id}{The id of the module, used to identify the UI elements.}
}
\description{
Creates the UI interface for the landing page module inside
the ClustAssess Shiny application.
}
\note{
This function should not be called directly, but in the context of the
app that is created using the \code{write_shiny_app} function.
}
