% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-sandbox.R
\name{ui_sandbox}
\alias{ui_sandbox}
\title{UI - Sandbox module}
\usage{
ui_sandbox(id)
}
\arguments{
\item{id}{The id of the module, used to identify the UI elements.}
}
\description{
Creates the UI interface for the sandbox module inside
the ClustAssess Shiny application.
}
\note{
This function should not be called directly, but in the context of the
app that is created using the \code{write_shiny_app} function.
}
