% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{calculate_markers_shiny}
\alias{calculate_markers_shiny}
\title{Calculate markers in the shiny context}
\usage{
calculate_markers_shiny(
  cells1,
  cells2,
  logfc_threshold = 0,
  min_pct_threshold = 0.1,
  average_expression_threshold = 0,
  min_diff_pct_threshold = -Inf,
  rank_matrix = NULL,
  feature_names = NULL,
  used_slot = "data",
  norm_method = "LogNormalize",
  pseudocount_use = 1,
  base = 2
)
}
\description{
to be completed
}
