% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC.R
\name{consensus_cluster}
\alias{consensus_cluster}
\title{Consensus Clustering and Proportion of Ambiguously Clustered Pairs}
\usage{
consensus_cluster(
  x,
  k_min = 3,
  k_max = 100,
  n_reps = 100,
  p_sample = 0.8,
  p_feature = 1,
  p_minkowski = 2,
  dist_method = "euclidean",
  linkage = "complete",
  lower_lim = 0.1,
  upper_lim = 0.9
)
}
\arguments{
\item{p_minkowski}{}
}
\value{

}
\description{
Calculate consensus clustering and proportion of ambiguously
clustered pairs (PAC)
}
