% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{element_agreement}
\alias{element_agreement}
\title{Element-Wise Average Agreement}
\usage{
element_agreement(reference_clustering, clustering_list)
}
\arguments{
\item{reference_clustering}{A Clustering objects for the reference clustering
that each clustering in clustering_list is compared to.}

\item{clustering_list}{A list of Clustering objects used to calculate
the element-wise average agreement}
}
\value{
a vector containing the element-wise average agreement.
}
\description{
Element-Wise Average Agreement
}
