% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{element_sim_elscore}
\alias{element_sim_elscore}
\title{The element-centric clustering similarity for each element.}
\usage{
element_sim_elscore(clustering1, clustering2)
}
\arguments{
\item{clustering1}{The first Clustering.}

\item{clustering2}{The second Clustering.}
}
\value{
The element-centric similarity between the two clusterings for each
element as a vector.
}
\description{
The element-centric clustering similarity for each element.
}
