% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{element_sim_matrix}
\alias{element_sim_matrix}
\title{Title}
\usage{
element_sim_matrix(clustering_list, output_type = "matrix")
}
\arguments{
\item{clustering_list}{A list of Clustering objects to be compared with
element-centric similarity.}

\item{output_type}{A string specifying whether the output should be a
matrix or a data.frame.}
}
\value{
a matrix or data.frame containing the pairwise ECS values.
}
\description{
Title
}
