% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-2-graph-construction.R
\name{get_highest_prune_param_embedding}
\alias{get_highest_prune_param_embedding}
\title{Relationship Between Number of Nearest Neighbors and Graph Connectivity}
\usage{
get_highest_prune_param_embedding(embedding, n_neigh)
}
\arguments{
\item{embedding}{An object or a concatenation of objects returned
by the `get_nn_conn_comps` method.}
}
\value{
A ggplot2 object with boxplots for the connected component distributions.
}
\description{
Display the distribution of the number connected components
obtained for each number of neighbors across random seeds.
}
\note{
The number of connected components is displayed on a logarithmic scale.
}
\examples{
set.seed(2021)

}
