% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{get_nn_conn_comps}
\alias{get_nn_conn_comps}
\title{Nearest Neighbors - Connected Components Link}
\usage{
get_nn_conn_comps(
  object,
  n_neigh_sequence,
  config_name = "",
  n_repetitions = 100,
  seed_sequence = NULL,
  graph_reduction_type = "UMAP",
  transpose = (graph_reduction_type == "PCA"),
  ncores = 1,
  ...
)
}
\arguments{
\item{object}{A matrix object. If the graph reduction type is PCA, the object
should be an expression matrix, with features on rows and observations on columns;
in the case of UMAP, the user could also provide a matrix associated to a PCA embedding.}

\item{n_neigh_sequence}{A sequence of the number of nearest neighbors.}

\item{config_name}{User specified string that uniquely describes the embedding characteristics.}

\item{n_repetitions}{The number of repetitions of applying the pipeline with different seeds; ignored if seed_sequence is provided by the user.}

\item{seed_sequence}{A custom seed sequence; if the value is NULL, the sequence will be built starting from 1 with a step of 100.}

\item{graph_reduction_type}{The graph reduction type, denoting if the graph should be built on either the PCA or the UMAP embedding.}

\item{transpose}{Decides whether the input object will be transposed or not.}

\item{ncores}{The number of parallel R instances that will run the code. If the value is set to 1, the code will be run sequentially.}

\item{...}{Additional arguments passed to the `irlba::irlba` or the `uwot::umap` method, depending on the value of graph_reduction_type.}
}
\value{
A list having one field associated with a number of nearest neighbors.
Each value contains an array of the number of connected components
obtained on the specified number of repetitions.
}
\description{
One of the steps in the PhenoGraph pipeline is building a graph
by using the kNN method on a reduced-space embedding. This method returns an
object that describes the relationship between different number of nearest
neighbors and the connectivity of the graph. In the context of graph clustering,
the number of connected components can be used as a
lower bound for the number of clusters. The calculations are performed multiple
times by changing the seed at each repetition.
}
\examples{
# get the PCA embedding of the data
pca_embedding = irlba::irlba(as.matrix(mtcars), nv = 2)
pca_embedding = pca_embedding$u \%*\% diag(pca_embedding$d)
rownames(pca_embedding) = rownames(mtcars)

# run the function on the PCA embedding
nn_conn_comps_obj = get_nn_conn_comps(object = pca_embedding,
    n_neigh_sequence = c(2,5,15),
    config_name = "mt_cars",
    n_repetitions = 1,
    graph_reduction_type = "UMAP",
    # the following parameters are used by the umap function and are not mandatory
    min_dist = 0.3,
    n_neighbors = 5,
    metric = "cosine")
}
