% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{merge_partitions}
\alias{merge_partitions}
\title{Merge Partitions}
\usage{
merge_partitions(
  partition_list,
  ecs_thresh = 1,
  order_logic = c("freq", "agreement", "none")
)
}
\arguments{
\item{partition_list}{A list of flat disjoint membership vectors.}

\item{ecs_thresh}{A numeric: the ecs threshold.}

\item{order_logic}{Variable indicating the method of ordering the partitions.
It can take these three values:
\itemize{
\item "freq": order the partitions based on their frequencies. The partition with
the highest frequency will be the first on the list (default).
\item "agreement": order the partitions based on their average agreement index.
The average agreement index of a partition is calculated as the mean of
the ECS scores between that partition and the other partitions from the list.
The partition with the highest agreement will be  the first on the list.
\item "none": do not perform any ordering (not recommended).
}}
}
\value{
a list of the merged partitions
}
\description{
Merge flat disjoint clusterings whose pairwise ECS score is
above a given threshold. The merging is done using a complete linkage approach.
}
\examples{
initial_list <- list(c(1, 1, 2), c(2, 2, 2), c("B", "B", "A"))
merge_partitions(initial_list, 1)
}
