% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{merge_resolutions}
\alias{merge_resolutions}
\title{Merge resolutions}
\usage{
merge_resolutions(res_obj, ncores = 1)
}
\arguments{
\item{res_obj}{A list associated to a configuration field from the object
returned by the \code{get_resolution_importance} method.}

\item{ncores}{The number of parallel R instances that will run the code.
If the value is set to 1, the code will be run sequentially.}
}
\value{
A list having one field assigned to each number of clusters. A number
of cluster will contain a list of all merged partitions. To avoid duplicates,
\code{merged_partitions} with threshold 1 is applied.
}
\description{
Merge partitions obtained with different resolution values
that have the same number of clusters.
}
\examples{
merge_resolutions(list(
 "0.8" = list(
   "2" = list(
     list(mb = c(1,1,1,2),
          freq = 2),
     list(mb = c(1,1,2,2),
          freq = 5)
   )
 ),
 "0.9" = list(
   "3" = list(
     list(mb = c(1,2,3,3),
          freq = 1)
   ),
   "2" = list(
     list(mb = c(2,2,3,3),
          freq = 2)
   )
 )
))
}
