% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC.R
\name{pac_convergence}
\alias{pac_convergence}
\title{PAC Convergence Plot}
\usage{
pac_convergence(pac_res, k_plot)
}
\arguments{
\item{k_plot}{}
}
\value{

}
\description{
Plot PAC across iterations for a set of k to assess convergence.
}
