% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{plot_clustering_difference_boxplot}
\alias{plot_clustering_difference_boxplot}
\title{Graphical representation of the clustering method importance}
\usage{
plot_clustering_difference_boxplot(clustering_difference_object, text_size = 3)
}
\arguments{
\item{clustering_difference_object}{An object returned by the
`get_clustering_difference_object` method.}

\item{text_size}{The size of the labels above boxplots.}
}
\value{
A ggplot object
}
\description{
Display, for each clustering method and each resolution value,
the distribution of the EC consistency as a boxplot. We will use the `filtered`
field of the object returned by the `get_clustering_difference_object` method.
Above each boxplot, the number of clusters will be displayed.
}
\examples{
adj_matrix = Seurat::FindNeighbors(as.matrix(mtcars),
    k.param = 5,
    nn.method = "rann",
    verbose = FALSE,
    compute.SNN = FALSE)$nn
clust_diff_obj = get_clustering_difference(graph_adjacency_matrix = adj_matrix,
    resolution = c(0.5, 1),
    n_repetitions = 1,
    algorithm = 1,
    verbose = FALSE)
plot_clustering_difference_boxplot(clust_diff_obj)
}
