% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{plot_clustering_difference_facet}
\alias{plot_clustering_difference_facet}
\title{Clustering Method Stability Facet Plot}
\usage{
plot_clustering_difference_facet(
  clustering_difference_object,
  embedding,
  low_limit = 0,
  high_limit = 1,
  grid = TRUE
)
}
\arguments{
\item{clustering_difference_object}{An object returned by the
`get_clustering_difference_object` method.}

\item{embedding}{An embedding (only the first two dimensions will be used for
visualisation).}

\item{low_limit}{The lowest value of ECC that will be displayed on the embedding.}

\item{high_limit}{The highest value of ECC that will be displayed on the embedding.}

\item{grid}{Boolean value indicating whether the facet should be a grid (where each
row is associated with a resolution value and each column with a clustering method) or
a wrap.}
}
\value{
A ggplot2 object.
}
\description{
Display the distribution of the EC consistency for each
clustering method and each resolution value on a given embedding The `all`
field of the object returned by the `get_clustering_difference_object` method is used.
}
\examples{
pca.mt=prcomp(as.matrix(mtcars))$x
adj_matrix = Seurat::FindNeighbors(pca.mt,
    k.param = 5,
    nn.method = "rann",
    verbose = FALSE,
    compute.SNN = FALSE)$nn
clust_diff_obj = get_clustering_difference(graph_adjacency_matrix = adj_matrix,
    resolution = c(0.5, 1),
    n_repetitions = 1,
    algorithm = 1,
    verbose = FALSE)
plot_clustering_difference_facet(clust_diff_obj, pca.mt[,1:2])
}
