% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-3-graph-clustering.R
\name{plot_clustering_overall_stability}
\alias{plot_clustering_overall_stability}
\title{Clustering Method Stability Boxplot}
\usage{
plot_clustering_overall_stability(
  clust_object,
  value_type = c("k", "resolution"),
  summary_function = median
)
}
\arguments{
\item{clust_object}{An object returned by the
`get_clustering_difference_object` method.}

\item{text_size}{The size of the labels above boxplots.}

\item{boxplot_width}{Used for adjusting the width of the boxplots; the value will
be passed to the `width` argument of the `ggplot2::geom_boxplot` method.}

\item{dodge_width}{Used for adjusting the horizontal position of the boxplot; the value
will be passed to the `width` argument of the `ggplot2::position_dodge` method.}
}
\value{
A ggplot2 object with the EC consistency distributions. Higher
consistency indicates a more stable clustering.
}
\description{
Display EC consistency across clustering method and resolution
values. The `filtered` field of the object returned by the
`get_clustering_difference_object` method is used.
Above each boxplot, the number of clusters is displayed.
}
\examples{
set.seed(2021)
# create an artificial PCA embedding
pca_embedding <- matrix(runif(100 * 30), nrow = 100)
rownames(pca_embedding) <- as.character(1:100)
colnames(pca_embedding) <- paste0("PCA_", 1:30)

adj_matrix <- Seurat::FindNeighbors(pca_embedding,
    k.param = 10,
    nn.method = "rann",
    verbose = FALSE,
    compute.SNN = FALSE
)$nn
clust_diff_obj <- assess_clustering_stability(
    graph_adjacency_matrix = adj_matrix,
    resolution = c(0.5, 1),
    n_repetitions = 10,
    algorithm = 1:2,
    verbose = FALSE
)
plot_clustering_overall_stability(clust_diff_obj)
}
