% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{plot_k_resolution_corresp}
\alias{plot_k_resolution_corresp}
\title{Correspondence between the resolution parameter and the number of clusters}
\usage{
plot_k_resolution_corresp(
  res_object_list,
  res_object_names = NULL,
  given_height = 0.7
)
}
\arguments{
\item{res_object_list}{An object or a concatenation of objects returned by the
`get_resolution_importance` method.}

\item{res_object_names}{Custom names that the user could assing to each
configuration; if not specified, the plot will use the generated configuration
names.}

\item{given_height}{Used for adjusting the vertical position of the boxplot; the value
will be passed in the `width` argument of the `ggplot::position_dodge` method.}
}
\value{
A ggplot object
}
\description{
For each configuration provided in the object, display what
number of clusters appear for different values of the resolution parameters.
Different shapes of points indicate different configuration, while the color
illustrates the frequency of the most common partition. The frequency is calculated
as the fraction between the number of total appearances and the number of
runs.
}
\examples{
# get the PCA embedding of the data
pca_embedding = irlba::irlba(as.matrix(mtcars), nv = 2)
pca_embedding = pca_embedding$u \%*\% diag(pca_embedding$d)
rownames(pca_embedding) = rownames(mtcars)

# run the function on the pca embedding
resolution_result = get_resolution_importance(embedding = pca_embedding,
   resolution = c(0.8, 1),
   n_neigh = c(5, 7),
   n_repetitions = 1,
   clustering_method = 1,
   graph_type = 2,
   object_name = "mt_cars")
plot_k_resolution_corresp(resolution_result)
}
