% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{plot_n_neigh_ecs}
\alias{plot_n_neigh_ecs}
\title{Graph construction parameteres - ECC facet}
\usage{
plot_n_neigh_ecs(nn_ecs_object)
}
\arguments{
\item{nn_ecs_object}{An object or a concatenation of objects returned by the
`get_nn_importance` method.}
}
\value{
A ggplot facet object
}
\description{
Display, for all configurations consisting in different number
of neighbors, graph types and base embeddings, the EC Consistency of the partitions
obtained over multiple runs on an UMAP embedding.
}
\examples{
nn_importance_obj = get_nn_importance(object = as.matrix(mtcars),
    n_neigh_sequence = c(2,5,15),
    n_repetitions = 1,
    graph_reduction_type = "PCA",
    algorithm = 1,
    transpose = FALSE, # the matrix is already observations x features, so we won't transpose it
    # the following parameter is used by the irlba function and is not mandatory
    nv = 2)
plot_n_neigh_ecs(nn_importance_obj)
}
