% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-based-parameter-assessment.R
\name{plot_n_neigh_k_correspondence}
\alias{plot_n_neigh_k_correspondence}
\title{Graphical representation of the link between NN and number of clusters}
\usage{
plot_n_neigh_k_correspondence(nn_object_n_clusters)
}
\arguments{
\item{nn_object_n_clusters}{An object or a concatenation of objects returned by the
`get_nn_importance` method.}
}
\value{
A ggplot object
}
\description{
Display, for each number of neighbors, the distribution of the
number of clusters obtained at different seeds. The distribution is
displayed as a boxplot.
}
\note{
The number of clusters is displayed on a logarithmic scale.
}
\examples{
nn_importance_obj = get_nn_importance(object = as.matrix(mtcars),
    n_neigh_sequence = c(2,5,15),
    n_repetitions = 1,
    graph_reduction_type = "PCA",
    algorithm = 1,
    transpose = FALSE, # the matrix is already observations x features, so we won't transpose it
    # the following parameter is used by the irlba function and is not mandatory
    nv = 2)
plot_n_neigh_k_correspondence(nn_importance_obj)
}
