% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{print,Clustering-method}
\alias{print,Clustering-method}
\title{Print an Object}
\usage{
\S4method{print}{Clustering}(x)
}
\arguments{
\item{x}{The Clustering object.}
}
\value{
The printed character string.
}
\description{
Prints out information about the Clustering, including
number of elements.
}
\keyword{internal}
