% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECS.R
\name{weighted_element_consistency}
\alias{weighted_element_consistency}
\title{Weighted Element-Centric Consistency}
\usage{
weighted_element_consistency(
  clustering_list,
  weights = NULL,
  calculate_sim_matrix = FALSE
)
}
\description{
calculate the element consistency of a list of clusterings where each object has a weight
}
