% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-app.R
\name{write_objects}
\alias{write_objects}
\title{Writing objects}
\usage{
write_objects(
  clustassess_object,
  expression_matrix,
  metadata,
  project_folder = ".",
  compression_level = 6,
  chunk_size = 100,
  gene_variance_threshold = 0,
  summary_function = median,
  qualpalr_colorspace = "pretty"
)
}
\description{
to be completed
}
