% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/shiny-app.R
\name{write_shiny_app}
\alias{write_shiny_app}
\alias{write_shiny_app.Seurat}
\alias{write_shiny_app.default}
\title{Writing the shiny app folder}
\usage{
write_shiny_app(object, ...)

\method{write_shiny_app}{Seurat}(
  object,
  assay_name,
  clustassess_object,
  project_folder,
  compression_level = 6,
  summary_function = median,
  shiny_app_title = "",
  organism_enrichment = "hsapiens",
  height_ratio = 0.6,
  qualpalr_colorspace = "pretty"
)

\method{write_shiny_app}{default}(
  object,
  metadata,
  clustassess_object,
  project_folder,
  compression_level = 6,
  summary_function = median,
  shiny_app_title = "",
  organism_enrichment = "hsapiens",
  height_ratio = 0.6,
  qualpalr_colorspace = "pretty"
)
}
\description{
to be completed

to be completed

to be completed
}
