% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-app.R
\name{write_shiny_app}
\alias{write_shiny_app}
\title{Writing the shiny app folder}
\usage{
write_shiny_app(
  seurat_object,
  assay_name,
  clustassess_object,
  project_folder,
  compression_level = 6,
  summary_function = median,
  qualpalr_colorspace = "pretty"
)
}
\description{
to be completed
}
